Move Beef.Rooms into access and mutations
