+++ kousa/lib/beef/queries/rooms.ex
@@ -1,0 +1,3 @@
+defmodule Beef.Queries.Rooms do
+
+end
